Designer: Add file watcher for automatically reloading Configurations view contents
Before this change, the configurations view content was loaded when the
configuration folder was selected and was never changed after it (except,
when the workspace changes). However, files may be created in the
configurations folders (like items, rules and so on), and it would be
great if the designer automatically recognizes these changes and loads
them into the Configuration file navigator.

This change adds the feature by adding a WatchService to the directories
once the configuration directory is selected. The watcher will
automatically reload the view, if files were created or removed (changes
doesn't make sense, as it wouldn't change the view at all).