Add new option to create a new file
For now, just a new file is possible, the user still has to type the
ending to the filename and has to select the configuration folder by its
own. In the future it would be possible to add seperate wizards for items,
rules, and so on which helps creating specific openHAB types.

This change also tries to fix some things regarding the project setup,
more will follow in the next commit.