Add new receivedCommandItem variable for rules
The receivedCommandItem varoable in the body of a rule allows to access
the item, which triggers the run of the current rule. Only available in
rules which uses the "received command" in the when clause. Could be used
to e.g. do something like this:

rule "Test"

when
	Item Test received command
	or Item Test2 received command
then
	receivedCommandItem.sendCommand(OFF)
end

Or anything what is possible with a named item, too.